refactor(server): extract IP header parsing into security core microcrate
