+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
   "crates/bitnet-eval-core",     # SRP: deterministic eval/scoring math helpers
   "crates/bitnet-feature-matrix",
   "crates/bitnet-server",
+  "crates/bitnet-server-security-core",
   "crates/bitnet-cli",
   "crates/bitnet-st2gguf",      # SafeTensors to GGUF converter
   "crates/bitnet-st-tools",     # SafeTensors inspection & merge utilities
@@ -118,6 +119,7 @@ default-members = [
   "crates/bitnet-sampling",
   "crates/bitnet-cli",
   "crates/bitnet-server",
+  "crates/bitnet-server-security-core",
   "crates/bitnet-st2gguf",
   # Testing infrastructure crates (always built/tested by default)
   "crates/bitnet-bdd-grid",
